// src/api/generated/api.ts
export interface ServerControllersModelsAuthUserDTO {
    'id'?: string;
    'username'?: string | null;
    'passwordHash'?: string | null;
    'role'?: ServerControllersModelsEnumsEnumAuth;
}


export interface ServerControllersModelsClientDTO {
    'id'?: string;
    'companyName'?: string | null;
    'phoneNumber'?: string | null;
}
// ...
export const ServerControllersModelsEnumsEnumAuth = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2
} as const;

export type ServerControllersModelsEnumsEnumAuth = typeof ServerControllersModelsEnumsEnumAuth[keyof typeof ServerControllersModelsEnumsEnumAuth];
// ...
export interface ServerControllersModelsJournalRecordDTO {
    'id'?: string;
    'plantId'?: string;
    'growthStageId'?: string;
    'employeeId'?: string;
    'plantHeight'?: number;
    'fruitCount'?: number;
    'condition'?: ServerControllersModelsEnumsEnumCondition;
    'date'?: string;
}


export interface ServerControllersModelsLoginRequestDto {
    'username'?: string | null;
    'password'?: string | null;
}
export interface ServerControllersModelsLoginResponseDto {
    'username'?: string | null;
    'token'?: string | null;
}
export interface ServerControllersModelsPlantDTO {
    'id'?: string;
    'clientId'?: string;
    'specie'?: string | null;
    'family'?: string | null;
    'flower'?: ServerControllersModelsEnumsEnumFlowers;
    'fruit'?: ServerControllersModelsEnumsEnumFruit;
    'reproduction'?: ServerControllersModelsEnumsEnumReproduction;
}


export interface ServerControllersModelsRegisterRequestDto {
    'email'?: string | null;
    'password'?: string | null;
}
export interface ServerControllersModelsSeedDTO {
    'id'?: string;
    'plantId'?: string;
    'maturity'?: string | null;
    'viability'?: ServerControllersModelsEnumsEnumViability;
    'lightRequirements'?: ServerControllersModelsEnumsEnumLight;
    'waterRequirements'?: string | null;
    'temperatureRequirements'?: number;
}


export interface ServerControllersModelsUpdateUserRoleRequestDto {
    'newRole'?: DomainModelsEnumsEnumAuth;
}


export interface ServerControllersModelsUserDTO {
    'id'?: string;
    'phoneNumber'?: string | null;
}
// ...
export interface AuthApiInterface {
    /**
     * 
     * @summary Аутентификация пользователя
     * @param {AuthApiApiV1AuthLoginPostRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1AuthLoginPost(requestParameters?: AuthApiApiV1AuthLoginPostRequest, options?: RawAxiosRequestConfig): AxiosPromise<ServerControllersModelsLoginResponseDto>;

    /**
     * 
     * @summary Регистрация нового пользователя
     * @param {AuthApiApiV1AuthRegisterPostRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1AuthRegisterPost(requestParameters?: AuthApiApiV1AuthRegisterPostRequest, options?: RawAxiosRequestConfig): AxiosPromise<ServerControllersModelsLoginResponseDto>;

}

/**
 * Request parameters for apiV1AuthLoginPost operation in AuthApi.
 */
export interface AuthApiApiV1AuthLoginPostRequest {
    readonly serverControllersModelsLoginRequestDto?: ServerControllersModelsLoginRequestDto
}

/**
 * Request parameters for apiV1AuthRegisterPost operation in AuthApi.
 */
export interface AuthApiApiV1AuthRegisterPostRequest {
    readonly serverControllersModelsRegisterRequestDto?: ServerControllersModelsRegisterRequestDto
}

/**
 * AuthApi - object-oriented interface
 */
export class AuthApi extends BaseAPI implements AuthApiInterface {
    /**
     * 
     * @summary Аутентификация пользователя
     * @param {AuthApiApiV1AuthLoginPostRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1AuthLoginPost(requestParameters: AuthApiApiV1AuthLoginPostRequest = {}, options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).apiV1AuthLoginPost(requestParameters.serverControllersModelsLoginRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Регистрация нового пользователя
     * @param {AuthApiApiV1AuthRegisterPostRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1AuthRegisterPost(requestParameters: AuthApiApiV1AuthRegisterPostRequest = {}, options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).apiV1AuthRegisterPost(requestParameters.serverControllersModelsRegisterRequestDto, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export interface ClientApiInterface {
    /**
     * 
     * @summary Изменить роль пользователя.
     * @param {ClientApiApiV1ClientsClientIdRolePatchRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1ClientsClientIdRolePatch(requestParameters: ClientApiApiV1ClientsClientIdRolePatchRequest, options?: RawAxiosRequestConfig): AxiosPromise<ServerControllersModelsAuthUserDTO>;

    /**
     * 
     * @summary Получить клиентов с возможностью фильтрации.
     * @param {ClientApiApiV1ClientsGetRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1ClientsGet(requestParameters?: ClientApiApiV1ClientsGetRequest, options?: RawAxiosRequestConfig): AxiosPromise<Array<ServerControllersModelsClientDTO>>;

    /**
     * 
     * @summary Получить клиента по идентификатору.
     * @param {ClientApiApiV1ClientsIdGetRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1ClientsIdGet(requestParameters: ClientApiApiV1ClientsIdGetRequest, options?: RawAxiosRequestConfig): AxiosPromise<ServerControllersModelsClientDTO>;

    /**
     * 
     * @summary Получить записи журнала клиента.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1ClientsJournalRecordsGet(options?: RawAxiosRequestConfig): AxiosPromise<Array<ServerControllersModelsJournalRecordDTO>>;

    /**
     * 
     * @summary Получить растения клиента.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1ClientsPlantsGet(options?: RawAxiosRequestConfig): AxiosPromise<Array<ServerControllersModelsPlantDTO>>;

}

/**
 * Request parameters for apiV1ClientsClientIdRolePatch operation in ClientApi.
 */
export interface ClientApiApiV1ClientsClientIdRolePatchRequest {
    /**
     * Идентификатор пользователя.
     */
    readonly clientId: string

    /**
     * Данные для изменения роли.
     */
    readonly serverControllersModelsUpdateUserRoleRequestDto?: ServerControllersModelsUpdateUserRoleRequestDto
}

/**
 * Request parameters for apiV1ClientsGet operation in ClientApi.
 */
export interface ClientApiApiV1ClientsGetRequest {
    /**
     * Название компании клиента (опционально).
     */
    readonly companyName?: string

    /**
     * Телефон клиента (опционально).
     */
    readonly phoneNumber?: string
}

/**
 * Request parameters for apiV1ClientsIdGet operation in ClientApi.
 */
export interface ClientApiApiV1ClientsIdGetRequest {
    /**
     * Идентификатор клиента.
     */
    readonly id: string
}

/**
 * ClientApi - object-oriented interface
 */
export class ClientApi extends BaseAPI implements ClientApiInterface {
    /**
     * 
     * @summary Изменить роль пользователя.
     * @param {ClientApiApiV1ClientsClientIdRolePatchRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1ClientsClientIdRolePatch(requestParameters: ClientApiApiV1ClientsClientIdRolePatchRequest, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiV1ClientsClientIdRolePatch(requestParameters.clientId, requestParameters.serverControllersModelsUpdateUserRoleRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Получить клиентов с возможностью фильтрации.
     * @param {ClientApiApiV1ClientsGetRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1ClientsGet(requestParameters: ClientApiApiV1ClientsGetRequest = {}, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiV1ClientsGet(requestParameters.companyName, requestParameters.phoneNumber, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Получить клиента по идентификатору.
     * @param {ClientApiApiV1ClientsIdGetRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1ClientsIdGet(requestParameters: ClientApiApiV1ClientsIdGetRequest, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiV1ClientsIdGet(requestParameters.id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Получить записи журнала клиента.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1ClientsJournalRecordsGet(options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiV1ClientsJournalRecordsGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Получить растения клиента.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1ClientsPlantsGet(options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiV1ClientsPlantsGet(options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export interface UserApiInterface {
    /**
     * 
     * @summary Получение списка всех пользователей (только для администраторов)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1UsersGet(options?: RawAxiosRequestConfig): AxiosPromise<Array<ServerControllersModelsAuthUserDTO>>;

    /**
     * 
     * @summary Получение информации о текущем пользователе
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    apiV1UsersMeGet(options?: RawAxiosRequestConfig): AxiosPromise<ServerControllersModelsUserDTO>;

}

/**
 * UserApi - object-oriented interface
 */
export class UserApi extends BaseAPI implements UserApiInterface {
    /**
     * 
     * @summary Получение списка всех пользователей (только для администраторов)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1UsersGet(options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).apiV1UsersGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Получение информации о текущем пользователе
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiV1UsersMeGet(options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).apiV1UsersMeGet(options).then((request) => request(this.axios, this.basePath));
    }
}